requirements.txt
openpyxl==3.1.5
python-dotenv==1.2.1
requests==2.32.5
tzdata==2025.2
